requirements.txt
Flask
requests
dlib
opencv-python-headless
numpy
google-cloud-vision
gunicorn